+++ src/App.js
@@ -3,12 +3,12 @@ import { Switch } from 'react-router-dom'
 import './less/index.less'
 import WrappedRegistrationForm from './components/register'
 import LoginForm from './components/login'
-import ResetPasswordForm from './components/resetPassword'
-import SetNewPasswordForm from './components/setNewPassword'
+import ResetPasswordForm from './components/ResetPassword/resetPassword'
+import SetNewPasswordForm from './components/ResetPassword/setNewPassword'
 import CreateStoreForm from './components/createStore/firstView'
 import AddLogoForm from './components/createStore/addLogo'
-import CreateItem from './components/CreateItem'
-import UpdateItem from './components/updateItem'
+import CreateItem from './components/Products/CreateItem'
+import UpdateItem from './components/Products/updateItem'
 import PrivateRoute from './components/Auth/PrivateRoute'
 import PublicRoute from './components/Auth/PublicRoute'
 import Main from './components/inventory'
