+++ src/components/Products/updateItem.js
@@ -1,25 +1,52 @@
 import React, { useState, useEffect } from 'react'
+import { connect, useDispatch, useSelector } from 'react-redux'
 import { Form, Input, Icon, Button, message, Upload, Spin } from 'antd'
 import axios from 'axios'
-import AxiosAuth from './Auth/axiosWithAuth'
-import history from '../history'
-import { connect, useSelector } from 'react-redux'
-import { setLoading, setErrors, clearErrors, getCurrentUser } from '../state/actionCreators'
-import useCurrency from './hooks/useCurrency'
+import AxiosAuth from '../Auth/axiosWithAuth'
+import history from '../../history'
+import { setLoading, setErrors, clearErrors } from '../../state/actionCreators'
+import * as creators from '../../state/actionCreators'
+import useCurrency from '../hooks/useCurrency'
 
-const productURL = 'https://shopping-cart-eu3.herokuapp.com/api/store/products'
-
-function CreateItem ({ dispatch, form, isLoading }) {
+function UpdateItem (props) {
+  const [item, setItem] = useState([])
   const [fileList, setFileList] = useState([])
   const [cloudList, setCloudList] = useState([])
   const { TextArea } = Input
+  const itemId = props.match.params.id
+  const productURL = `https://shopping-cart-eu3.herokuapp.com/api/store/products/${itemId}`
+
+  const dispatch = useDispatch()
+
+  useEffect(() => {
+    dispatch(creators.getCurrentUser())
+  }, [dispatch])
+
+  const currencyDescription = useSelector(state => state.user.user.currency)
+
+  const sign = useCurrency(currencyDescription)
+
+  useEffect(() => {
+    AxiosAuth()
+      .get(
+        `https://shopping-cart-eu3.herokuapp.com/api/store/products/${itemId}`
+      )
+      .then(res => {
+        const newFileList = res.data.images.map((url, idx) => ({
+          uid: -idx,
+          name: `photo ${idx}.jpg`,
+          url
+        }))
+        setFileList(newFileList)
+        setItem(res.data)
+      })
+  }, [itemId])
+
   const handleChange = info => {
     let fileList = [...info.fileList]
-
     // 1. Limit the number of uploaded files
     // Only to show two recent uploaded files, and old ones will be replaced by the new
     fileList = fileList.slice(-4)
-
     // 2. Read from response and show file link
     fileList = fileList.map(file => {
       if (file.response) {
@@ -28,17 +55,9 @@ function CreateItem ({ dispatch, form, isLoading }) {
       }
       return file
     })
-
+    // const newFile= fileList
     setFileList(fileList)
   }
-
-  useEffect(() => {
-    dispatch(getCurrentUser())
-  }, [dispatch])
-
-  const currencyDescription = useSelector(state => state.user.user.currency)
-
-  const sign = useCurrency(currencyDescription)
 
   const dummyRequest = ({ file, onSuccess }) => {
     const image = new FormData()
@@ -59,37 +78,37 @@ function CreateItem ({ dispatch, form, isLoading }) {
     }, 0)
   }
 
-  useEffect(() => {
-    dispatch(setLoading(false))
-  }, [dispatch])
-
   const handleSubmit = e => {
     e.preventDefault()
-    form.validateFieldsAndScroll((err, values) => {
+    props.form.validateFieldsAndScroll((err, values) => {
+      const images = [
+        ...cloudList,
+        ...fileList.filter(image => image.url).map(image => image.url)
+      ]
       const payload = {
         name: values.name,
         description: values.description,
         price: values.price,
         stock: values.stock || 0,
-        images: cloudList
+        images
       }
       if (!err) {
-        dispatch(setLoading(true))
+        props.dispatch(setLoading(true))
         AxiosAuth()
-          .post(productURL, payload)
+          .put(productURL, payload)
           .then(res => {
-            message.success('Item Added')
-            dispatch(setLoading(false))
-            dispatch(clearErrors())
+            message.success('item updated')
+            props.dispatch(setLoading(false))
+            props.dispatch(clearErrors())
             history.push('/inventory')
           })
           .catch(error => {
-            dispatch(setLoading(false))
-            dispatch(setErrors(error.response.data))
+            props.dispatch(setLoading(false))
+            props.dispatch(setErrors(error.response.data))
             message.error(Object.values(error.response.data)[0])
           })
       } else {
-        message.error('Enter Required Fields')
+        message.error('Validation failed')
       }
     })
   }
@@ -99,8 +118,7 @@ function CreateItem ({ dispatch, form, isLoading }) {
     history.push('/inventory')
   }
 
-  const { getFieldDecorator } = form
-
+  const { getFieldDecorator } = props.form
   const formItemLayout = {
     labelCol: {
       xs: { span: 24 },
@@ -124,19 +142,15 @@ function CreateItem ({ dispatch, form, isLoading }) {
     }
   }
 
-  const createItemComponent = (
-    <Spin spinning={isLoading}>
+  return (
+    <Spin spinning={props.isLoading}>
       <div className='cover' id='createUpdate'>
         <div id='header'>
-          <h2 id='get-started'>
-            Upload new
-            <br />
-            store item
-          </h2>
+          <h2 id='get-started'>Update {item.name}</h2>
         </div>
         <div id='uploadHead' style={{ height: '30%', width: '100%' }}>
           <Upload
-            style={{ height: '20%', width: '20%' }}
+            // style={{ height: '20%', width: '20%' }}
             listType='picture-card'
             fileList={fileList}
             customRequest={dummyRequest}
@@ -148,6 +162,7 @@ function CreateItem ({ dispatch, form, isLoading }) {
         <Form className='inputForm' {...formItemLayout} onSubmit={handleSubmit}>
           <Form.Item>
             {getFieldDecorator('name', {
+              initialValue: item.name,
               rules: [
                 {
                   message: 'Name'
@@ -159,32 +174,19 @@ function CreateItem ({ dispatch, form, isLoading }) {
               ]
             })(<Input placeholder='Name' />)}
           </Form.Item>
-
           <Form.Item>
             {getFieldDecorator('price', {
-              rules: [
-                {
-                  message: 'Enter a price'
-                },
-                {
-                  required: true,
-                  message: 'Enter a price'
-                }
-              ]
+              initialValue: item.price
             })(<Input placeholder='Price' addonBefore={sign} />)}
           </Form.Item>
-
           <Form.Item>
             {getFieldDecorator('stock', {
-              rules: [
-                {
-                  message: 'Enter stock'
-                }
-              ]
+              initialValue: item.stock
             })(<Input placeholder='Stock' />)}
           </Form.Item>
           <Form.Item>
             {getFieldDecorator('description', {
+              initialValue: item.description,
               rules: [
                 {
                   message: 'Enter a description'
@@ -196,7 +198,6 @@ function CreateItem ({ dispatch, form, isLoading }) {
               ]
             })(<TextArea placeholder='Description' allowClear />)}
           </Form.Item>
-
           <Form.Item {...tailFormItemLayout}>
             <Button type='primary' htmlType='submit'>
               Done
@@ -209,15 +210,11 @@ function CreateItem ({ dispatch, form, isLoading }) {
       </div>
     </Spin>
   )
-
-  return createItemComponent
 }
-
-const CreateItemForm = Form.create({ name: 'createItem' })(CreateItem)
-
+const UpdateItemForm = Form.create()(UpdateItem)
 const mapStateToProps = state => ({
   isLoading: state.user.isLoading,
   errors: state.user.errors
 })
 
-export default connect(mapStateToProps, null)(CreateItemForm)
+export default connect(mapStateToProps, null)(UpdateItemForm)
